refactor twotower models.
