+++ demo/twotowers/dssm/conf/dssm_batch_neg_ml_1m.yaml
@@ -14,10 +14,10 @@ test_path: ${MY_S3_BUCKET}/movielens/1m/negsample/num_negs_10/test.parquet
 item_path: ${MY_S3_BUCKET}/movielens/1m/negsample/prob/item.parquet
 
 ## model configuration
-two_tower_module_class: python.algos.twotowers.TwoTowerBatchNegativeSamplingModule
-two_tower_agent_class: python.algos.twotowers.TwoTowerBatchNegativeSamplingAgent
+two_tower_module_class: python.algos.twotowers.dssm.TwoTowerBatchNegativeSamplingModule
+two_tower_agent_class: python.algos.twotowers.dssm.TwoTowerBatchNegativeSamplingAgent
 two_tower_estimator_class: metaspore.TwoTowerRetrievalEstimator
-# two_tower_estimator_class: python.two_tower_retrieval_milvus.TwoTowerRetrievalEstimator
+# two_tower_estimator_class: python.twotower.TwoTowerRetrievalEstimator
 
 user_column_name: ${MY_S3_BUCKET}/movielens/1m/schema/dssm/user_column_schema
 user_combine_schema: ${MY_S3_BUCKET}/movielens/1m/schema/dssm/user_combine_schema
