--- python/algos/twotowers/__init__.py
@@ -1,2 +1,0 @@
-from .dssm_net import UserModule, ItemModule, SimilarityModule
-from .batch_negative_sampling_agent import TwoTowerBatchNegativeSamplingModule, TwoTowerBatchNegativeSamplingAgent
